restructured templates files
