+++ dlp/snippets/Templates/templates_test.py
@@ -17,9 +17,12 @@ import uuid
 
 import google.api_core.exceptions
 import google.cloud.storage
-import pytest
 
-import templates
+import create_inspect_template as ct
+import delete_inspect_template as dt
+import list_inspect_templates as lt
+
+import pytest
 
 
 UNIQUE_STRING = str(uuid.uuid4()).split("-")[0]
@@ -29,20 +32,20 @@ TEST_TEMPLATE_ID = "test-template" + UNIQUE_STRING
 
 def test_create_list_and_delete_template(capsys: pytest.CaptureFixture) -> None:
     try:
-        templates.create_inspect_template(
+        ct.create_inspect_template(
             GCLOUD_PROJECT,
             ["FIRST_NAME", "EMAIL_ADDRESS", "PHONE_NUMBER"],
             template_id=TEST_TEMPLATE_ID,
         )
     except google.api_core.exceptions.InvalidArgument:
         # Template already exists, perhaps due to a previous interrupted test.
-        templates.delete_inspect_template(GCLOUD_PROJECT, TEST_TEMPLATE_ID)
+        dt.delete_inspect_template(GCLOUD_PROJECT, TEST_TEMPLATE_ID)
 
         out, _ = capsys.readouterr()
         assert TEST_TEMPLATE_ID in out
 
         # Try again and move on.
-        templates.create_inspect_template(
+        ct.create_inspect_template(
             GCLOUD_PROJECT,
             ["FIRST_NAME", "EMAIL_ADDRESS", "PHONE_NUMBER"],
             template_id=TEST_TEMPLATE_ID,
@@ -51,12 +54,12 @@ def test_create_list_and_delete_template(capsys: pytest.CaptureFixture) -> None:
     out, _ = capsys.readouterr()
     assert TEST_TEMPLATE_ID in out
 
-    templates.list_inspect_templates(GCLOUD_PROJECT)
+    lt.list_inspect_templates(GCLOUD_PROJECT)
 
     out, _ = capsys.readouterr()
     assert TEST_TEMPLATE_ID in out
 
-    templates.delete_inspect_template(GCLOUD_PROJECT, TEST_TEMPLATE_ID)
+    dt.delete_inspect_template(GCLOUD_PROJECT, TEST_TEMPLATE_ID)
 
     out, _ = capsys.readouterr()
     assert TEST_TEMPLATE_ID in out
